requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.32
psycopg[binary,pool]==3.2.1

pydantic==2.9.0
httpx==0.27.2

google-cloud-tasks==2.16.5

web3==6.20.1

python-dotenv==1.0.1
